requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
chromadb==0.4.22
sentence-transformers==2.2.2
APScheduler==3.10.4
torch==2.1.2
numpy==1.26.3